Move jackson cbor decoder in a dedicated package
