+++ leshan-core/src/main/java/org/eclipse/leshan/senml/cbor/jackson/SenMLCbor.java
@@ -12,7 +12,7 @@
  *     Boya Zhang - initial API and implementation
  *******************************************************************************/
 
-package org.eclipse.leshan.senml.cbor;
+package org.eclipse.leshan.senml.cbor.jackson;
 
 import org.eclipse.leshan.senml.SenMLPack;
